Fix & move channel configuration in firmware
The channel request handling code in gp.c was bogus. It tried to reply
only when the 'next channel' matched our radio channel. Instead it
should have replied to each Channel Request received, ignoring the
value of 'next channel' received.

During the first stage of commissioning, the GPD sends Channel Requests
on all possible channels (11-26). If we receive one, it obviously means
the GPD is already talking on our channel. It will open its RX window
on that same channel before switching and trying the 'next channel'.

Since the reply should be sent in time, it was moved in the firmware.