powertagd: fix a race during network creation
When running the create command, a race condition seems to happen
when the NCP is still bringing the network up as we try to send
the GP key. The NCP then sends a NAK frame which is currently not
well handled in our ASH implementation.

Here's a debug log of the issue:

ZigBee network successfully created!
    channel:  26
    TX power: +0 dBm
    PAN id:   0x0ec8
[ DBG] ezsp: sending EZSP_CUSTOM_FRAME frame (18 bytes):
[ DBG] ezsp:   [ 11 0F 80 71 03 71 DF A1 72 3C EA 8A 38 28 7A 7E ]
[ DBG] ezsp:   [ E1 48 ]
[ DBG] ash: sending data frame 5
[ DBG] ash: got data frame 5:
[ DBG] ash:   [ 3C 90 01 19 00 90 ]
[ DBG] ash: got data frame (6 bytes)
[ DBG] ash: sending ACK for frame 6
[ DBG] ezsp: got callback frame EZSP_STACK_STATUS_HANDLER
[ DBG] ezsp: got EZSP_STACK_STATUS_HANDLER frame (1 bytes):
[ DBG] ezsp:   [ 90 ]
[ DBG] ezsp: queueing callback frame EZSP_STACK_STATUS_HANDLER
[ DBG] ash: got NAK frame
[WARN] ash: got NAK but no data was sent, ignoring
[ ERR] ezsp: no reply to EZSP_CUSTOM_FRAME

As a temporary workaround, waiting a little bit just after network
creation is enough.